Three small fixes (#110)
* Fix exit codes for connect server

I think we've got this backwards - we should `os.Exit(1)` if
`ListenAndServeTLS` returns something _other_ than
`http.ErrServerClosed`.

* Dry up cmd/client/main.go

We're currently running the same tests for compressed and uncompressed
clients. I don't think that'll change, so let's DRY up the main package
so we're less likely to make a mistake later on.

* Consume iterators idiomatically

Fix the Connect implementation of `StreamingInputCall` to consume the
iterator more idiomatically.

Co-authored-by: Doria Keung <doria.keung@gmail.com>