Add insecure flag to client CLI commands (#470)
This adds an `insecure` flag to the client CLI command args. It allows
for starting a client over cleartext on HTTP/1.1 and HTTP/2.

If `insecure` is set, then all Connect HTTP/1.1 and HTTP/2
implementations will use an `http://` scheme and cleartext. Note that if
`insecure` is set, then `cert` and `key` must not be.

Also, the `insecure` flag is not allowed with HTTP/3 implementations.

A sample command using this flag for an insecure client:

`go run cmd/client/main.go --port 3000 --implementation connect-h1
--insecure`