Add connect-node-fastify server (#477)
This adds a new Connect server using `connect-fastify` and
`connect-node`. It wires it into the docker-compose setup and allows for
serving HTTP/1.1 and HTTP/2 traffic using cleartext and TLS. Fastify
does not offer HTTP/3 support as of yet, so that path is not available.
Note: this does not yet use TLS, but can be added as a follow-up PR.

In addition, this also refactors a few things such as:
* adds TypeScript support to the web portion
* renames images, Dockerfiles, etc. to match the language so that adding
new languages can follow one pattern. For example,
`client-connect-to-server-connect-go-h2` becomes
`client-connect-go-to-server-connect-go-h2` and uses
`Dockerfile.crosstestgo` instead of `Dockerfile.crosstest`.
* updates to connect-go v1.9.0 for
https://github.com/bufbuild/connect-go/pull/528 which fixes an issue
with parsing response headers from a connect-node server.
* updates to connect-es v0.11.0 for
https://github.com/bufbuild/connect-es/pull/695, which fixes an issue
with parsing response bodies for non-JSON content types.