Pin cpp plugin version and fix minor README issues (#554)
This fixes a few minor issues:

* Pins the C++ plugin version to `v23.4` since leaving it unspecified
was always using the latest, which is now v24.x and has breaking changes
in the generated code involving things like TSAN and others (at least
with the Bazel setup in the repo). Since v24.x is Editions, it is
probably best to pin to a major version back for now. This will fix the
failing CI badge for `conformance-cc`
* README improvements
  * Fixes broken links
  * Renames `connect.build` links to be `connectrpc.com`
  * Updates clients and servers descriptions
  * Minor grammar fixes
* Fixes paths in tests to reflect `connectrpc/conformance/v1` paths.