V2: Load test cases from embedded test suite data (#619)
This change adds the logic for embedding test suite data and for loading the
data, filtering it according to active config cases, and then grouping
by server process. In this case, a "server process" corresponds to a
`ServerCompatRequest` and thus a single server process that will be
started by the test runner.