V2: Add test result handler (#622)
This adds a `testResults` type that accumulates the results of a
conformance test run and can report the results at the end. This also
adds a `knownFailingTrie`, which is a prefix tree that is an index of
all known-failing test cases. This is used to invert the expectation of
a test: if a test case is configured as known-failing, the conformance
test runner accepts when the case results in an error but will complain
if the test case actually passes (since it means the known-failing
config is stale).