V2: Update client and server programs to use delimited streaming input (#626)
1. The biggest change is that both of these now read/write delimited
messages. This shouldn't be necessary on the server, since it only ever
reads and writes one message. However, in practice, I ran into a lot of
difficulty consuming all input from the server's stdout: it seems that
the `os/exec` package doesn't correctly close the associated stdout
writer when it sees EOF on the process's stdout. So trying to use
`io.ReadAll` to fully consume the process's stdout just hung forever.

In any event, it seemed like a reasonable change for consistency, so
both client and server can be implemented in the same way, except that
the server does not need to loop over the input to EOF but just needs to
read one message. Also, there's no risk of the runner just hanging if
the server-under-test fails to close its stdout. And it solves the
problem because the length prefix means the runner knows exactly how
many bytes to read for the server reply, instead of waiting until it
sees EOF.
   
2. This also changes the flags management in the reference client and
server. They were previously ignoring the args slice passed to them,
which means they would always be looking at the flags passed to the test
runner when they were spun up in-process. While making this change, I
also update these programs so that they won't start if you pass spurious
positional command-line args.

3. I've updated the server to include the observed timeout in the
payload response. This currently looks at the timeout on the context,
but it should probably be changed to instead dig it out of headers. (But
that's a bigger pain since the logic is a little different depending on
whether the protocol in use is connect or grpc.)

4. Finally, I've updated the server test files to have `.json`
extensions and updated the client test files to use consistent ports.
The client tests now use port 23456 for HTTP/1 and 23457 for HTTP/2.