V2: Smarter assertions about whether the client response matches the expected response (#629)
This addresses the TODO about how comparing the actual response to
expected needed to be more sophisticated.

In particular:
1. It is acceptable for the actual headers/trailers to contain more than
what's in the test expectation. Not only is it common for some HTTP
client/server frameworks to add add'l headers (like "Date",
"User-Agent", etc.), but we also don't want to hard-code all of the
protocol-specific headers into the test case data (e.g. "Content-Type").
2. It is acceptable for a client or server to merge headers and trailers
for responses that have empty response streams and an error. These can
be conveyed via "trailers-only responses" in the gRPC protocol, in which
case it is not possible to separate headers and trailers. Furthermore,
such responses are actually "headers-only" on the wire, so we allow the
sets to be merged into either headers or trailers under these specific
conditions.
3. It is acceptable for the server to omit the Connect GET query params
in its response.
4. It is acceptable for the client to omit the raw Connect error
details.
5. `Any` messages need smarter treatment than just `proto.Equal` since
two semantically equivalent messages could be serialized differently,
which means the `value` byte slices in two messages may differ and cause
`proto.Equal` to consider them unequal.