Add initial implementation for gRPC server (#628)
To gain confidence in our conformance tests and to ensure parity with
gRPC, we will be running tests against a pure gRPC server (i.e.
grpc-go). The server, like the Connect reference server, implements the
`ConformanceService` endpoints and behaves in the same way. This PR
implements the `Unary` endpoint with the others to follow.

The intent is that the gRPC server (and eventual client) won't stick
around forever, so rather than mix the gRPC-specific code (such as
dealing with `Metadata`, gRPC errors, and gRPC codecs), the code has
been rearranged so that the gRPC logic can live in its own directories
and be removed easily.