V2: Add TLS and HTTP/3 support to reference server (#647)
This adds some helpers to the `internal` package for dealing with TLS
configuration. This also updates the reference server to use it. FWIW, I
don't think we need to bother with adding this to the grpc server -- it
shouldn't be doing anything materially different with TLS to be worth
adding that.