V2: Add Bidi test cases (#648)
This adds two test cases for bidi streaming, both testing successful
cases. The two cases cover full and half duplex streams.

Note this also adds a few make targets to build the binaries and to run
the tests. It also fixes a minor oversight in the bidi endpoint that was
not setting response headers or trailers.

To run the tests:

`make runservertests`

---------

Co-authored-by: Joshua Humphries <2035234+jhump@users.noreply.github.com>