V2: Add basic success and error tests (#655)
This adds basic success and error tests for the remaining rpc types
(unary, server stream, and client stream (error))

Note that the error tests for unary and client stream had to be divided
by protocol (Connect vs. gRPC), since the gRPC tests return a
trailers-only response, which requires the `expectedResponse` to be
structured differently for Connect vs. gRPC.