V2: Autopopulate expected responses (#667)
This autopopulates the expected response information based on what is
provided for requests in the test yaml files. This allows for
eliminating a lot of boilerplate in the yaml files for straight-forward
tests for all stream types.

The logic looks at the request information and then dynamically builds a
`ClientResponseResult` based on what is being sent to the server.

This also allows for overriding the autopopulation by explicitly
specifying an `expectedResult` in the yaml file. This can be helpful for
one-off or out-of-the-ordinary requests, such as the Unimplemented call
or evaluating gRPC response trailers.

Note that the above could also be built into the autopopulation, but
this base path might be a better place to start to gauge whether this is
a good idea.