V1: (Last?) tweak to the protos (#685)
These are hopefully final updates to the protos before pushing to a new
BSR module and merging v1 to main.

These will be used in the changes to the reference client and server to
use raw requests and responses (preview in #682). They include a
backwards-incompatible change (making the length unsigned), so I wanted
to get this in before pushing to the BSR. After we push to the BSR, we
can enable `buf breaking` checks to keep us from making incompatible
changes going forward.