V1: Use sigterm instead of interrupt to stop a process (#683)
If trying to use `go run ...` as the server under test, the test runner
would hang. It turns out that this is because `go run ...` ignores
SIGINT, which is how the runner was previously trying to stop a test
process.

Now it uses SIGTERM. Also, it now cancels the process's context.

This also improves the output when using `-v`. Previously, it would log
information about all server instances up-front. Now it logs as it goes,
so it's possible to see its progress (if it makes any). This also tweaks
some of the messages, so it also shows when it's using the normal
reference client or server vs. the gRPC version.