V1: The grpc-go module uses raw values for binary headers, not base64-encoded (#689)
Unlike connect-go, the grpc-go library expects raw binary data in a
`metadata.MD` for binary headers (those whose key ends in "-bin"). This
was meant to be for convenience. But it means we have to be mindful when
converting to/from the `Header` proto, which needs them to be encoded.