Preparing for a v1.0 (#703)
This merges the long-lived `v1` branch to main, replacing the interop
cross-tests on main with an entirely new model that is data-driven. This
lets us add test cases to this repo, in a declarative format (YAML
files), and they can then easily be run against all implementations
(without having to re-implement the test case in each client).

Instead of docker containers, needed to ship multiple implementations
with differing technology stacks, we now need only a single,
statically-linked binary that is part of each release (and downloadable
via GitHub release artifacts). This binary functions as a test runner
and includes a reference client and reference server linked in, for
evaluating other implementations.

New implementations can use the conformance tests by implementing a
client and/or server for the ConformanceService defined in the protos herein
(https://buf.build/connectrpc/conformance). These will be data-driven
programs that communicate with the test runner by receiving
configuration/inputs via stdin and writing status/results to stdout.

---------

Co-authored-by: Steve Ayers <sayers@buf.build>