Add 'buf push' and 'make checkrelease' to CI and remove refs to 'v1' branch (#714)
This also re-organizes the Buf checks in their own file, following the
layout of some other repos, such as bufbuild/buf. This applies the same
recipe as in that repo so that the Buf actions are only run when the
protos actually change and so that the new "push" task is only run for
commits on `main`.