Clarify rules around response definitions (#711)
This tries to clarify the rules around response definitions and how the
server should behave in various scenarios where response definition is
either not set at all or is partially set with various values. To
complement this, it also adds tests for each RPC type that have
partially-set response definitions as well as no response definition.
Numerous areas in the servers had to be fixed to check for a `nil`
response definition and to act accordingly.

The protos have been updated to reflect the rules now and should
(hopefully) be clearer and more thorough. The basic rules are:

* for unary responses, if no response definition is set, then return the
request info and an empty response
* for server responses, if no response definition is set (or is set, but
no responses are desired), do not send anything on the stream and either
return without error or return an error if defined.

The bidi flow has changed slightly. The basic premise for full and half
duplex is 'if you have no responses to send, check to see if an error
was provided. If so, throw that. Otherwise, just do nothing and return
without error'. As a result, it is now possible for a full duplex call
to have less responses desired than requests sent. In these cases, the
flow should follow the above premise.

Also, the `connect-only` and `grpc-only` test files have been merged
into `basic` since those should be covered the same for all protocols
now.

Note: this replaces #696 since there were way too many merge conflicts
with the merge of v1 into main. This should be identical to PR (just
based off of main and with no merge conflicts).

---------

Co-authored-by: Joshua Humphries <2035234+jhump@users.noreply.github.com>
Co-authored-by: Chris Pine <cpine@buf.build>
Co-authored-by: Nick Snyder <nickdsnyder@gmail.com>