Fix some error-handling issues in the reference client's server stream impl (#718)
There were two cases where the `serverStream` method was returning
the wrong kind of error and failing to return a response proto that described
the error. That was causing server-stream timeout errors to frequently be
interpreted as a test case failure, even though the error was "deadline exceeded"
as expected.