Add test case for runner and remove races in I/O (#725)
This is a follow-up to #720.

Since that PR is merged, accessing the `outWriter` and `errWriter` from
multiple goroutines is racy. Even if the underlying writer were
thread-safe, this means that writing a message or line could end up
being "interrupted" by a concurrent write, interleaving lines and
leaving the output unreadable. This commit fixes that by
introducing a `Printer` interface with a thread-safe implementation
to do logging.

To make sure there aren't any actual data races in this code, this also
adds a new test that exercises a more significant portion of the
conformance test runner, including the concurrent aspects of the
runner. So running `go test -race` can now verify that these aspects
are race-free.