Fix cancellation and timeout failures (#734)
This makes various improvements to fix the failing cancellation and
timeout tests:

* Removes `ctx.Err()` checks in the clients since this can be an
ambiguous error on the client and could potentially be unrelated to the
error we received from the RPC framework.
* Adds logic to immediately send response headers on server and bidi
streams so that clients can be unblocked reading this data. Previously
this was causing clients to block on reading this data and was
preventing them from cancelling at the intended time.
* Adds docs to the protos specifying the above.
* Adds cancellation logic to gRPC `BidiStream`
* Configures some tests with appropriate delays so that the cancellation
is properly communicated.

All of the tests in `bidi-cancellation.yaml` and `bidi-timeouts.yaml`
have been folded into the `basic.yaml` suite of tests since these apply
to and pass all permutations.