Add flags to test runner allowing user to supply TLS cert+key, port, and bind address for in-process reference server(s) (#735)
In certain environments, it may not be possible to configure trust roots
and use self-signed certificates. Also, it may be necessary to
know/configure the listen port so that the reference server can be
properly exported for use by an external client. For these reasons, this
adds flags to the test runner that allow the user to supply a TLS cert
and key and to configure the port and bind address on which reference
server(s) listen. Without these, the servers will generate a self-signed
cert+key and will listen on "127.0.0.1:0".