Add error tests (#736)
This adds new tests for each Connect error code for both unary and
streaming responses, since errors are represented differently on the
wire for each. The tests verify that each are able to be returned and
that the client is able to determine the error.

To test unary responses, the `Unary` endpoint was used and for streaming
responses, the `ServerStream` endpoint is used.

In addition, all the previous error test cases for specific RPC types
were moved to the new `errors.yaml` file.