Support sending raw HTTP request in reference client (#737)
This still expects the test case to include a normal request, too. That
allows us to continue using the "expected response" computation, based
on the request. But we can then override the actual HTTP request, to
simulate some behavior (mainly around on-the-wire encoding) that the
reference client doesn't usually exhibit. The main things that are _not_
possible with raw requests are the ability to do cancellations or to
introduce delays between writes.