Minor improvements related to populating expected responses (#738)
Populating expected responses in test cases should be the responsibility
of the test case library. So instead of making the caller do this, it's
now done as part of computing the test cases.

Also, this allows the automatic population of expected responses, even
if a raw request or response is used. This is because the normal request
and response definition will also be used alongside the raw request or
response. So if the expected response from that is still correct, just
auto-populate with that. So we only need to provide an explicit expected
response if the raw request or response causes the server to behave
differently.