Refactor `TestPopulateExpectedResponse` tests (#742)
This PR attempts to make some improvements to the
`TestPopulateExpectedResponse` test cases. Currently, they are very hard
to follow and add to due to how messages are constructed ahead of time
and then set in the `testCase` list. It is confusing to look at an
individual test and figure out what's going on without bouncing around
to the various definitions.

The reason this was originally coded this way is because a vast majority
of these messages needed converted to an `Any` and `anypb.New` returns a
message and an error, so that error needed verified. To alleviate this
and help remove these pre-constructed messages, a new function has been
created: `asAnySlice`, which takes a variable amount of messages and
converts them to a slice of `Any` messages. It also verifies the error
itself so that the callsites can now just accept a single value.

Note that this PR does not add any additional tests and it admittedly
does not shrink the size of the `TestPopulateExpectedResponse` test.
But, each test case should now be self-contained and easier to grok by
just looking at the `testCase.request` and `testCase.expected` values
without having to bounce all over the place. It should also be much
easier to add new tests without having to pre-construct a multitude of
message definitions.