Support sending raw HTTP response in reference server (#740)
This one's a little more complicated than the client side because of how
we have to use middleware to hook into the HTTP framework's response
writer. We rely on a Connect interceptor to examine the request message
and store a raw response, if so configured, into a placeholder in context.
The middleware uses the absence or presence of that value to decide
whether to respect the method calls made by the handler (for sending
a normal response) or to ignore them and then send the raw response
at the end. Because it waits until the end to send the raw response, it
cannot work with full-duplex bidi streams. But it can work with half-duplex
bidi streams.