Reference client and server should support half-duplex bidi over HTTP 1.1 (#741)
If some other implementation under test wants to enable half-duplex bidi
over HTTP 1.1. (which is ostensibly supported by the configuration
flags), then we need the reference client and server to actually work
that way, too.

Out of the box, connect-go refuses to support bidi streams over HTTP
1.1, because they _might_ be full-duplex, which can't work over HTTP 1.1
(and otherwise can result in inexplicable errors that may look like
network partitions or other connection stability issues since they
manifest as the client's upload being cancelled as soon as the server
starts its reply). So we have to "trick" the connect-go runtime in order
to support this. Also, it means that the actual bidi stream handler in
the server impl cannot send back headers immediately since, in
half-duplex operation, that immediately cancels the client's upload.