Test runner needs to add synthetic headers to raw request, too (#758)
These are extra headers used by the reference server, to verify that the
client created the request in the expected way. They were previously
only added to the structured request that normal clients use. But they
also need to be included in the raw request, for test cases that use
raw requests with the reference client.

This also makes it so that the test case name header is _always_ added,
even when not running against the reference server. This is useful if
tracing through and debugging an implementation -- you can easily tell
which test case the request is for because it is always in the request
headers.