Add command-line flags for more control over test cases to run and how they're handled (#757)
This adds a few new features:
* The existing `--test-file` flag can now be specified multiple times,
so one can run multiple suites at the same time, directly from a YAML
file instead of from the embedded test case data.
* The existing `--known-failing` flag is changing to be a pattern, and
can be specified multiple times. To refer to a file of patterns (old
behavior), prefix the value with "@".
* The `--known-flaky` flag, which indicates patterns of test case names
that are known to be flaky. The difference between "known failing" and
"known flaky" is that if a known flaky test happens to succeed, the
conformance tests will still succeed. With a known failing test,
however, if it passes, that causes the run to fail. The "known failing"
set should generally be preferred so that if/when a test case is fixed,
the conformance tests will enforce that you also update your list of
known failing test cases.
* The `--run` flag, which indicates patterns of test case names to
include in the run. If not specified, _all_ test cases that apply (per
config) are included. This allows running just a single test suite, test
case, or test case permutation, depending on what the user is trying to
debug.
* The `--skip` flag, which is the reverse above. This indicates patterns
of test case names to _not_ bother running. In general, users should
prefer `--known-failing` or `--known-flaky` flags to opt-out of test
cases. But this is a convenient way to simply _not run_ particular test
cases when running the conformance tests.