Add `--trace` option to print full trace of HTTP requests and responses for failed test cases (#761)
This is a bit of a bear -- most of the complexity is in the minimal
amount of parsing that the middleware must do on the wire format, to
report relevant information about it in the trace and then to also
capture the end-of-stream message in the Connect streaming and gRPC-web
protocols.

So most of the code is in the new `tracer` internal package, and that's
where all of the gnarly bits are, too. The integrations into the
command-line and into the reference client and server are pretty
straight-forward.

This does not instrument the grpc-go reference client and server
implementations. We could add that later. It's relatively
straight-forward with the server, since it _can_ be configured to use
`net/http` (though it is not currently configured this way). But the
grpc-go client cannot be configured that way. So to collect traces, we
must point the grpc-go client at a local (in-process) HTTP proxy, that
can record the traces and then forward the requests/responses to the
real backend. This seems like enough of a headache that we can punt on
it and maybe revisit if people really need traces against the grpc-go
implementations.