Add wire tracer capabilities to reference client (#768)
This piggybacks off of the tracer framework to add wire tracing to the
reference client. Broadly, this implements the `Collector` interface and
intercepts the `Complete` function invocation to sniff the built trace
and acquire wire details.

The details tracked are:
- Actual HTTP status code
- Actual trailers on the wire
- Connect Raw Error Details (if applicable)

Note that the tests to verify raw error details trailers are not yet in
place to keep the complexity of this PR lower. They will be added in a
follow-up PR.

---------

Co-authored-by: Joshua Humphries <2035234+jhump@users.noreply.github.com>