Add checks in referenceclient to validate raw gRPC-Web trailers and actual HTTP trailers (#780)
This adds handling of the HTTP trailers and gRPC-Web end-stream messages
to the reference client. When in reference mode, it will examine them
and make sure they are appropriate, using stricter checks than
connect-go (so we can't just rely on the base connect-go client
implementation in the reference client to raise these issues in the
conformance results).