Add strict checks to reference client to validate a server's connect error and end stream JSON (#790)
This makes the checks significantly more strict than the base connect-go
implementation, in order to report a number of possible issues in server
JSON responses:
1. This will detect duplicate keys in JSON objects/maps. Without an
explicit check, "encoding/json" ignores duplicate keys, letting later
values for the same key overwrite earlier ones.
2. This ensures that all keys use the correct case. Sadly, even with
explicit field names in struct tags, "encoding/json" is still lenient
and uses a case-insensitive match for field names.
3. This double-checks the JSON type of each value. There are a few ways
in which "encoding/json" sometimes allows undesired types -- mainly with
the use of a JSON null for map, slice, and pointer fields. So this
performs an extra check of the actual types in the raw JSON.
4. This makes sure that the keys and values in the end-stream "metadata"
map are valid HTTP header names and values.
5. This also makes sure that if a "debug" key is present in an error
detail, that the message it describes actually matches the message
described in the binary data of the "value" key.