Prevent errant 'content-length: 0' request header from being printed (#815)
Previously, the reference client's trace output could print an errant
"Content-Length: 0" request header. It was incorrectly interpreting
the meaning of the `http.Request.ContentLength` field. It turns out
that the client's rules for when the header is added are complicated,
and it can also transparently add other headers. So in order to
_accurately_ show the real headers sent to the server, we need to
use the `httptrace` package.

There are a handful of other small changes in here, because I saw this
errant content-length issue when examining the output of tests run
against a grpc-java server. The conformance server implementation was
incorrectly leaving `Host` empty in the `ServerCompatResponse`, and yet
everything worked (except that the trace printed the URL without a host
name). I've changed it to explicitly use "127.0.0.1" as a default if the
server does not provide a host. And I also added some "TE: Trailers"
request headers to the gRPC raw requests in test cases, because
grpc-java servers complain in the logs if the requests do not include
this. (All proper gRPC clients _do_ include it.)