Add tracing support to gRPC impls by sniffing HTTP/2 frames on underlying TCP socket (#823)
This adds a new tracing capability -- to wrap a `net.Conn` that
represents an HTTP/2 connection and examine all of the HTTP/2 traffic to
assemble a trace. This is used to instrument a gRPC client and server to
support tracing, since they include their own HTTP/2 implementation
instead of using `net/http`.