Troubleshoot and fix issues in grpc-web JS conformance client (#831)
There were a couple of bugs in the conformance client that were
causing some tests to outright hang (promise was never resolved)
and some cases to have an incorrect result sent back to the test
runner (causing some tests that were actually correct/passing to
be misinterpreted as failing).

This adds some logging to the client so that any uncaught errors
in the browser can be logged to stderr (instead of being trapped
in the headless browser's JS console). It also adds a safeguard to
prevent hanging in the future: if a test takes longer than 15s, the
client will mark it as a failure and stop waiting for the promise to
be resolved.