Server response must include a certificate when using TLS (#832)
The test runner did not previously verify that the server returned a
cert when the request indicated that TLS was to be used. This empty
cert in the response would result in the reference client not using
TLS. This basically meant that TLS test cases could pass even when
TLS wasn't actually even used on the wire. The solution is pretty to
have the runner actually verify that the server provides a certificate.