Fix issues with test runner hanging (#829)
There were two main sources of the test runner freezing:

1. If the program under test spawned a child process, then even though the
child process may terminate, the grandchild process is keeping stdin, stdout,
and/or stderr open. The `os/exec` package also needs I/O goroutines to
complete before reporting the process completed, so hangs waiting for
these file descriptors to be closed. Note that the `os/exec` package has
some affordances for better handling this, in the form of the `Cmd`'s
special `StdinPipe()`, `StdoutPipe()`, and `StderrPipe()` methods. But
when I tried out using them, it was actually a bit _more_ code and
resulted in some of the child process's output getting truncated. So it
was not really an improvement. Fixing this involves abandoning the
child process if it won't exit and can't be promptly terminated.

2. We have a 20 second timeout for reading a response from the client.
But there was no such timeout for reading a response from the server. So
if something went awry, the test runner could hang indefinitely, awaiting
the server's response. We now apply a timeout to the server response.
We also now apply a size limit to both server and client response. This
limit means the most significant byte of the 32-bit length prefix should
always be zero, which should make it easier to distinguish inadvertent
text emitted to stdout from a length prefix for an extremely large response.