Relax expectation for unspecified errors, fix expectation for cardinality violations (#833)
There are numerous error conditions in gRPC where the actual behavior of
a client or server in the face of such an error is unspecified. In
particular, there is no specification for how a client or server should
react to an invalid/unexpected HTTP method or content-type. So this adds
a mechanism to supply additional acceptable error codes.

One that thing that **is** specified, but was codified with the wrong
expectation in these tests, is related to cardinality violations. A
cardinality violation is when exactly one request or response is
expected (for a unary RPC or the non-streaming half of a non-bidi
streaming RPC), but there is actually zero or more than one. This is
specified in a table that accompanies the docs for gRPC
codes. Search for "cardinality violation" in this page:
   https://grpc.github.io/grpc/core/md_doc_statuscodes.html