Reference and gRPC servers can send independent header and trailer metadata (#840)
The connect-go API doesn't provide a handler a direct way to send
independent response headers and trailers from a unary or client-stream
RPC. This is because a handler may return either a `*connect.Response`
or a `*connect.Error`, but not both. But a `*connect.Error` does not
allow setting both headers and trailers: it has only a `Meta()`
accessor, which ostensibly could be serialized using either headers or
trailers.

The reason this is the case is because frameworks can sometimes decide
to send all metadata together, such as in a "trailers-only" response
(allowed by the gRPC and gRPC-Web protocols for more compact replies
when there is no message data). Because a framework decision could
change whether a piece of metadata shows up as a header or a trailer,
clients should generally provide a combined view of headers and trailers
in the face of an error, to avoid bugs where the client looks for the
metadata as a header, but it was actually serialized as a trailer (i.e.
in a "trailers-only" response).

Having said all that, it can be surprising for users to see headers and
trailers combined in the response from the reference server, in an HTTP
trace. So to make the reference server's behavior more intuitive, we
send headers and trailers separately -- by going _around_ the direct
APIs offered by connect-go.

For client streams, there is an indirect way provided: one can access
the underlying `connect.StreamingHandlerConn` and then directly interact
with response headers and trailers.

But for unary RPCs, we have to be sneakier. For the Connect unary
protocol, all metadata gets serialized as headers. And trailers are
differentiated via a "trailer-" prefix. So we can just prefix the
metadata that is _intended_ to be trailers with "trailer-". For the gRPC
and gRPC-Web protocols, there are no simple hacks to get the connect-go
framework to reply as desired. So we leverage the existing "raw
response" functionality to go _around_ the framework and directly
control what is written to the wire for the responses for these cases.